src/requirements.txt
sodapy
google-cloud-storage
google-cloud-bigquery
pyarrow
pandas
pandas-gbq